🔥 Shereフォルダ削除
