+++ src/pages/login/Sign.tsx
@@ -1,4 +1,4 @@
-import { Button } from 'src/components/share/Button'
+import { Button } from 'src/components/Button'
 import { useSign } from 'src/hooks/useSign'
 
 export const Sign = () => {
