--- src/types/index.ts
@@ -1,3 +1,0 @@
-import type { AdminFormProps } from './types'
-
-export type { AdminFormProps }
