--- src/types/types.ts
@@ -1,8 +1,0 @@
-export type AdminFormProps = {
-  theme: string
-  name: string
-  date: string
-  time: string
-  link: string
-  discription: string
-}
